# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fincava — Colombian Agricultural B2B Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List verified suppliers
      parameters:
        - name: companyType
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - name: certification
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /supplier/profile:
    get:
      operationId: getMySupplierProfile
      tags: [suppliers]
      summary: Get authenticated supplier's own profile
      responses:
        "200":
          description: Supplier profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetail"

    patch:
      operationId: updateSupplierProfile
      tags: [suppliers]
      summary: Update supplier profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetail"

  # INQUIRIES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        supplierName:
          type: string
        supplierVerified:
          type: boolean
        supplierLogoUrl:
          type: ["string", "null"]
        name:
          type: string
        category:
          type: string
        subCategory:
          type: ["string", "null"]
        description:
          type: string
        origin:
          type: string
        altitude:
          type: ["string", "null"]
        process:
          type: ["string", "null"]
        variety:
          type: ["string", "null"]
        minOrderKg:
          type: number
        maxOrderKg:
          type: ["number", "null"]
        pricePerKgUSD:
          type: number
        availableKg:
          type: number
        harvestSeason:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        cupping:
          type: ["number", "null"]
        active:
          type: boolean
        featured:
          type: boolean
        avgRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - companyId
        - supplierName
        - supplierVerified
        - name
        - category
        - description
        - origin
        - minOrderKg
        - pricePerKgUSD
        - availableKg
        - images
        - certifications
        - active
        - featured
        - reviewCount
        - createdAt
...
    Certification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        issuer:
          type: string
        expiryDate:
          type: ["string", "null"]
        documentUrl:
          type: ["string", "null"]
        verified:
          type: boolean
      required:
        - id
        - type
        - issuer
        - verified
    Supplier:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          type: string
        country:
          type: string
        region:
          type: ["string", "null"]
        description:
          type: string
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        verified:
          type: boolean
        certifications:
          type: array
          items:
            type: string
        productCategories:
          type: array
          items:
            type: string
        productCount:
          type: integer
        avgRating:
          type: ["number", "null"]
        memberSince:
          type: string
      required:
        - id
        - userId
        - name
        - type
        - country
        - description
        - verified
        - certifications
        - productCategories
        - productCount
        - memberSince
    SupplierDetail:
      allOf:
        - $ref: "#/components/schemas/Supplier"
        - type: object
          properties:
            certificationDetails:
              type: array
              items:
                $ref: "#/components/schemas/Certification"
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            originStory:
              type: ["string", "null"]
            farmerName:
              type: ["string", "null"]
          required:
            - certificationDetails
            - products
            - reviews
    UpdateSupplierBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        country:
          type: string
        region:
          type: ["string", "null"]
        description:
          type: string
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        originStory:
          type: ["string", "null"]
        farmerName:
          type: ["string", "null"]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorCountry:
          type: ["string", "null"]
        productId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - authorId
        - authorName
        - productId
        - rating
        - verified
        - createdAt
...